pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "airflow-provider-kaggle"
authors = [
    {name = "Panagiotis Simakis", email = "sp1thas@autistici.org"},
]
license = {text = "Apache License 2.0"}
description = "A Kaggle Apache Airflow provider package."
classifiers = [
    "Framework :: Apache Airflow",
    "Framework :: Apache Airflow :: Provider",
    "Development Status :: 3 - Alpha",
]
dynamic = ["version"]
requires-python = "~=3.8"
dependencies = ["apache-airflow>=2.0", "kaggle>=1.5"]

[project.optional-dependencies]
dev = ["pre-commit"]
test = ["pytest"]

[project.urls]
Homepage = "https://github.com/sp1thas/airflow-provider-kaggle/"
Source = "https://github.com/sp1thas/airflow-provider-kaggle/"

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.entry-points.apache_airflow_provider]
provider_info = "kaggle_provider.__init__:get_provider_info"

[tool.setuptools]
include-package-data = false

[tool.setuptools.dynamic]
version = {attr = "kaggle_provider.__version__"}

[tool.setuptools.packages.find]
exclude = ["*tests.*", "*tests"]
